projects/python/8.fastapi-alembic/requirements.txt
fastapi[standard]
sqlmodel
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4
alembic
